+++ cli_options/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+* Support grouping options in the docs by section. You can now specify the `:doc_section`
+to any option, and pass a `:sections` option in the `CliOptions.docs/2` function for the
+headers and extended docs of each section.
+
 ## [v0.1.1](https://github.com/sportradar/elixir-workspace/tree/cli_options/v0.1.1) (2024-07-04)
 
 ### Added